requirements.txt
streamlit
pandas
plotly
supabase
google-generativeai
diskcache
langchain-community
